requirements.txt
flask
requests
tensorflow
transformers
numpy
python-dotenv
google-generativeai